==> R/00_dependencias.R <==
install.packages(c(
    "httr2",
    "readxl",
    "rbcb",
    "sidrar",
    "dplyr",
    "lubridate",
    "tidyr",
    "arrow",
    "tzdb",
    "tsibble",
    "feasts",
    "fabletools",
    "forecast",
    "modeltime.resample",
    "modeltime",
    "parsnip",
    "ranger",
    "tune",
    "modeltime.ensemble",
    "workflows",
    "timetk",
    "recipes",
    "rsample",
    "purrr",
    "ggplot2",
    "readr"
))
